Forced Competitor.fromDocSnap to get int eloScore, switched to CustomScrollView on CategoryPage, fixed streamRivalryVotes to use untyped map

+++ lib/models/competitor.dart
@@ -31,5 +31,5 @@ class Competitor extends Equatable {
   Competitor.fromDocumentSnapshot(DocumentSnapshot doc)
       : id = doc.id,
         item = Item.fromMap(doc.get("item")),
-        eloScore = doc.get("eloScore");
+        eloScore = (doc.get("eloScore")).toInt();
 }
